+++ medical-supplies-website/src/main/java/com/example/medicalsupplieswebsite/service/IPositionService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface IPositionService {
     List<Position> findAll();
-    Position findById(Long id);
+    Position findById(Integer id);
 }
